+++ worker/src/i18n/type.ts
@@ -47,6 +47,7 @@ export type LocaleMessages = {
     NewPasswordRequiredMsg: string
     InvalidAddressTokenMsg: string
     FailedUpdatePasswordMsg: string
+    CredentialLoginDisabledMsg: string
     PasswordLoginDisabledMsg: string
     EmailPasswordRequiredMsg: string
     AddressNotFoundMsg: string